fixed image views

+++ kerinci-internal/src/main/java/ch/szclsb/kerinci/internal/renderpass/KrcAttachmentDescription.java
@@ -1,18 +1,14 @@
 package ch.szclsb.kerinci.internal.renderpass;
 
-import ch.szclsb.kerinci.api.VkAttachmentDescription;
-
-import java.lang.foreign.MemorySegment;
-
 public record KrcAttachmentDescription(
         int flags,
         int format,
         int samples,
         int loadOp,
-        long storeOp,
+        int storeOp,
         int stencilLoadOp,
-        long stencilStoreOp,
+        int stencilStoreOp,
         int initialLayout,
-        long finalLayout
+        int finalLayout
 ) {
 }
